autenticacao
autenticacao

+++ pythonProject/controllers/usuario_controller.py
@@ -1,10 +1,12 @@
+from flask_login import current_user, login_required, login_user, logout_user
 from sqlalchemy.sql.functions import user
 
 from main import app
-from flask import render_template, request, jsonify, redirect, url_for
+from flask import flash, render_template, request, jsonify, redirect, url_for
 from models.User import *
 from models.Conexao import *
 from datetime import datetime
+import bcrypt
 @app.route("/usuario/create")
 def inserir():
     return "chegou na rota"
@@ -18,6 +20,7 @@ def login():
     return render_template("usuario/sign-in.html")
 
 @app.route('/dash', methods=['POST','GET'])
+@login_required
 def dash():
     if request.method == 'POST':
         login = request.form['login']
@@ -32,6 +35,7 @@ def dash():
 
 
 @app.route('/users', methods=['GET'])
+@login_required
 def users():
     # abrir a conexao
     SessionLocal = sessionmaker(autocommit=False, autoflush=False,
@@ -61,6 +65,7 @@ def save_user():
         # Adiciona o novo usuário ao banco de dados
         db.add(new_user)
         db.commit()
+        login_user(new_user)
         return render_template("dashboard.html")
     except Exception as e:
         print(e)
@@ -86,10 +91,12 @@ def login_do():
         return render_template("usuario/sign-in.html", msg="Usuário não existe")
     else:
         #usuário existe e eu vou verificar se a senha está valida.
-        if userDB.senha != senha:
-            return render_template("usuario/sign-in.html", msg="Senha inválida")
+        if bcrypt.checkpw(senha.encode('utf-8'),userDB.senha.encode('utf_8')):
+            login_user(userDB, remember=True)
+            return render_template("/dashboard.html")
         else:
-           return render_template("/dashboard.html")
+            return render_template("usuario/sign-in.html", msg="Senha inválida")
+           
 
 #usando aplicao restful
 @app.route('/login/do/restful', methods=['POST'])
@@ -118,6 +125,7 @@ def login_do_restful():
         return jsonify({"status": "erro", "message": str(e)}), 400
 
 @app.route('/users/delete/<user_id>', methods=['GET'])
+@login_required
 def delete_user(user_id):
    # abre a conexao
     SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
@@ -126,6 +134,10 @@ def delete_user(user_id):
     user = db.query(User).filter(User.id == user_id).first()
     if user is None:
         return "Usuário nao encontrado", 404
+    
+    if user.id == current_user.id:
+        return "Usuário logado nao pode ser deletado", 404
+    
     db.delete(user)
     db.commit()
 
@@ -135,6 +147,7 @@ def delete_user(user_id):
     return redirect(url_for('users'))
 
 @app.route('/users/pre-update/<user_id>', methods=['GET'])
+@login_required
 def pre_update_user(user_id):
     SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
     db = SessionLocal()
@@ -145,6 +158,7 @@ def pre_update_user(user_id):
     return render_template("/usuario/update.html", user=user)
 
 @app.route('/users/update/', methods=['POST'])
+@login_required
 def update_user():
    #obtendo os dados que o usuário digitou
     user_id = int(request.form['user_id'])
@@ -170,3 +184,12 @@ def update_user():
     db.close()
 
     return redirect(url_for('users'))
+@app.route("/logout")
+def logout():
+    logout_user()
+    return redirect(url_for('login'))
+
+@app.route("/profile")
+@login_required
+def profile():
+    return render_template("usuario/update.html",user=current_user);
